pkg/sanity: use different volume names
Within the test suite, each name is now unique. That simplifies
debugging and when one test fails to remove its volume, the other
tests are not affected. It also allows running tests in parallel.

The names are not generated dynamically, because that would make
debugging harder again (new names for each test run). The downside is
that volumes may have to get cleaned up outside of the test suite if
a test fails to delete them.

Fixes: kubernetes-csi/csi-test#89.