v3: ensure that dep can import v3.0.0
These symlinks ensure that "dep" can still import v3.0.0. Without
them, importing fails when it encounters the
github.com/kubernetes-csi/csi-test/v3 because there's no code under
that path without the symlinks. Symlinks work for files, but not for
directories (probably due to a check in dep which expects directories
for packages).

Upstream suggests to move the code itself ("Major subdirectory",
https://github.com/golang/go/wiki/Modules#releasing-modules-v2-or-higher). But
that messes with git history. Symlinks are simpler, but have the
disadvantage of only working on Unix. Users on Windows will have to
use "go mod" for importing.